+++ index.js
@@ -1,5 +1,6 @@
 import generateMessages from './generateMessages'
 import Toucan from 'toucan-js'
+import customTexts from './data/customTexts'
 
 function postToDiscord({ msg = '', channel = 'skills', embeds = [] }) {
   const channels = {
@@ -14,9 +15,54 @@ function postToDiscord({ msg = '', channel = 'skills', embeds = [] }) {
   })
 }
 
+async function getNonParticipants() {
+  const options = {
+    method: 'GET',
+    headers: {
+      Accept: 'application/json',
+      'x-api-user': X_API_USER,
+      'x-api-key': X_API_KEY,
+    },
+  }
+
+  const [rawQuest, rawParty] = await Promise.all(
+    fetch('https://habitica.com/api/v3/groups/party', options),
+    fetch('https://habitica.com/api/v3/groups/party/members', options)
+  )
+
+  const quest = await rawQuest.json()
+  const party = await rawParty.json()
+  const questMembers = Object.keys(quest.data.quest.members)
+  const partyMembers = party.data
+
+  return partyMembers
+    .filter((member) => questMembers.includes(member._id))
+    .map((nonParticipant) => {
+      return {
+        title: nonParticipant.profile.name,
+        url: 'https://habitica.com/profile/' + nonParticipant._id,
+      }
+    })
+}
+
 async function handleRequest(request, sentry) {
   try {
     const payload = await request.json()
+
+    if (
+      payload.webhookType === 'questActivity' &&
+      payload.type === 'questStarted'
+    ) {
+      const nonParticipants = await getNonParticipants()
+
+      await postToDiscord({
+        msg: customTexts.quest_non_participants,
+        channel: 'quest_non_participants',
+        embeds: nonParticipants,
+      })
+      return new Response('OK')
+    }
+
     const messages = await generateMessages(payload)
     for (const message of messages) {
       await postToDiscord(message)
